requirements.txt
streamlit
pandas
mlflow
python-dotenv
hopsworks
altair